Delta versioon 2.5.1.11
Selle versiooniga tulevad kaasa järgmised parandused ja täiendused:
128205 EDHS-230 Kuupäevavahemiku alguse ja lõpu reeglid 
131078 EDHS-295 Ddoc-i sees xls fail ei avane ekraanivormilt 
145017 EDHS-394 Muuta teade "Enne jätkamist paranda vead" 
157979 EDHS-448 Kaastäitja tööülesande teavitus läks teise asutusse 
162256 EDHS-508 Sarja õigustesse ei saa lisada dokumendihaldurid gruppi 
153829 SIM109 Postipoisist imporditud dokumentide pealkirja muutmine 
153907 SIM122 Asendaja määramine ja asenduse kuupäevade kontroll 
153910 SIM123 Nimekirja sorteerimise meelde jätmine 
153914 SIM129 Otsingu funktsionaalsuse täiendused 
153920 SIM132 Klassifikaatori väärtuste järjestamine 
153921 SIM134 Tühja lisaadressaadi rea mitte kandmine väljasaatmise kuvale 
160068 SIM137, EDHS-476 Delegeerimise nuppude joondus 
160181 SIM137, EDHS-485 Terviktöövoog on lahku löödud töövood blokis, ühes blokis olevad kasutajad on töövood blokis teises järjestuses kui oli käivitamisel 
159905, 149896, 159906 SIM138 Otsingute optimeerimine; otsingu tulemuste ekraanile tagasi pöördumise optimeerimine

Teadaolev viga: Delta uuendamisel versioonilt 1.10 (või Viljandi MV’s versioonilt 2.1) versioonile 2.5 ei tööta SIM087 raames realiseeritud skript (eemaldab olemasolevatelt puhkusekorraldustelt välja Mall). Antud skripti paranduse tarnime järgmise parandustarnega.

Delta uuendamisel versioonilt 1.10 (või Viljandi MV’s versioonilt 2.1) versioonile 2.5 eemaldatakse menetluses menüüpunkti alt kõik sissetulevad kirjad, millel pole käivitatud ühtegi terviktöövoogu ja mis on registreeritud varem kui eile (varem kui versiooni paigaldamisele eelneval kuupäeval) – selleks lisatakse ja käivitatakse neil üks tühi terviktöövoog.

Delta uuendamisel versioonilt 1.10 (või Viljandi MV’s versioonilt 2.1) versioonile 2.5 eemaldatakse saatmata menüüpunkti alt kõik lepingud, mis on registreeritud varem kui eile (varem kui versiooni paigaldamisele eelneval kuupäeval) – selleks lisatakse neile üks väljasaatmise kirje.

----------
Juhised süsteemiadministraatoritele:

Delta uuendamisel versioonilt 1.10 (või Viljandi MV’s versioonilt 2.1) versioonile 2.5
1) Liigutada glassfish/domains/DOMAIN/config kaustast (Viljandi MV’s Tomcat puhul tomcat kaustast) kõik *.csv failid Delta andmekausta (dir.root seadega määratud kaust)
2) alfesco-global.properties failis
2.1) lisada rida
         updater.version110To25Combined=true
2.2) kui 1.10 -> 2.5 uuendusel tahetakse testimiseks uuendada ainult 2 kuu andmed (nt kui testkeskkonnas jookseb skript aeglaselt), siis lisada rida
         updater.limitForTesting=true
2.3) kui vastavalt lisaarenduste nõudele 109 tahetakse muuta ka PP imporditud dokumentide pealkirju (SiM'is tahetakse, SMIT'is ja Viljandi MV's mitte), siis lisada read
         postipoissImportDocumentsUpdater.enabled=true
         postipoissImportDocumentsUpdater.importStartDate=01.01.2000 00:00:00
         postipoissImportDocumentsUpdater.importEndDate=06.09.2010 23:59:59
       ja määrata ajavahemik, milllal toimus PP andmete importimine
2.4) kui eksisteerib, siis eemaldada seade updater.archive.enabled, seda enam ei kasutata

Kui vanast ajast on jäänud glassfish/domains/DOMAIN/lib kausta postgresql-....jar fail, siis selle võib kustutada, see fail on olemas Delta WAR’i sees.


git-svn-id: https://svn.smit/tarkvara/edhs/trunk@102648 5fddf92a-9921-11dd-8d32-b999f448c456

+++ delta/source/java/ee/webmedia/alfresco/document/web/DocumentDialog.java
@@ -47,6 +47,7 @@ import org.springframework.web.jsf.FacesContextUtils;
 import ee.webmedia.alfresco.cases.model.CaseModel;
 import ee.webmedia.alfresco.classificator.enums.DocumentStatus;
 import ee.webmedia.alfresco.common.listener.RefreshEventListener;
+import ee.webmedia.alfresco.common.web.BeanHelper;
 import ee.webmedia.alfresco.common.web.ClearStateNotificationHandler;
 import ee.webmedia.alfresco.document.associations.model.DocAssocInfo;
 import ee.webmedia.alfresco.document.associations.web.AssocsBlockBean;
@@ -111,6 +112,7 @@ public class DocumentDialog extends BaseDialogBean implements ClearStateNotifica
     private boolean isDraft;
     private boolean showDocsAndCasesAssocs;
     private boolean skipInit;
+    private boolean docReloadDisabled;
 
     private Node node;
     private List<NodeRef> newInvoiceDocuments = new ArrayList<NodeRef>();
@@ -271,7 +273,11 @@ public class DocumentDialog extends BaseDialogBean implements ClearStateNotifica
     public void deleteDocument(@SuppressWarnings("unused") ActionEvent event) {
         Assert.notNull(node, "No current document");
         try {
+            validatePermission(node, DocumentCommonModel.Privileges.DELETE_DOCUMENT_META_DATA);
             getDocumentService().deleteDocument(node.getNodeRef());
+        } catch (UnableToPerformException e) {
+            MessageUtil.addStatusMessage(e);
+            return;
         } catch (AccessDeniedException e) {
             MessageUtil.addErrorMessage(FacesContext.getCurrentInstance(), "document_delete_error_accessDenied");
             return;
@@ -403,6 +409,7 @@ public class DocumentDialog extends BaseDialogBean implements ClearStateNotifica
     @Override
     public void init(Map<String, String> params) {
         validatePermissions();
+        BeanHelper.getVisitedDocumentsBean().getVisitedDocuments().add(node.getNodeRef());
         if (skipInit) {
             searchBlockBean.init(node, isIncomingInvoice());
             skipInit = false;
@@ -444,8 +451,13 @@ public class DocumentDialog extends BaseDialogBean implements ClearStateNotifica
     public void restored() {
         try {
             final boolean snapshotRestored = restoreSnapshot();
+            BeanHelper.getVisitedDocumentsBean().getVisitedDocuments().add(node.getNodeRef());
             if (!snapshotRestored) {
-                reloadDocAndClearPropertySheet();
+                if (!docReloadDisabled) {
+                    reloadDocAndClearPropertySheet();
+                } else {
+                    docReloadDisabled = false;
+                }
                 fileBlockBean.restore();
                 sendOutBlockBean.restore();
                 assocsBlockBean.restore();
@@ -467,12 +479,13 @@ public class DocumentDialog extends BaseDialogBean implements ClearStateNotifica
                 /** It's possible to change the type of the node that came from DVK */
                 getDocumentService().changeType(node);
             }
-            metadataBlockBean.save(isDraft, newInvoiceDocuments);
-            transactionsBlockBean.save();
+            if (transactionsBlockBean.save()) {
+                metadataBlockBean.save(isDraft, newInvoiceDocuments);
+                notifyModeChanged();
+            }
             logBlockBean.restore();
             isDraft = false;
             isFinished = false;
-            notifyModeChanged();
             ((MenuBean) FacesHelper.getManagedBean(context, MenuBean.BEAN_NAME)).processTaskItems(); // Update UserWorkingDocuments number
             return null;
         }
@@ -737,6 +750,11 @@ public class DocumentDialog extends BaseDialogBean implements ClearStateNotifica
             node = getDocumentService().getDocument(nodeRef);
             setupAction(false);
         }
+    }
+
+    public void addFile(@SuppressWarnings("unused") ActionEvent event) {
+        docReloadDisabled = true;
+        BeanHelper.getAddFileDialog().start(null);
     }
 
     @Override
